refactor: split DB view into two components
