+++ log-viewer/modules/components/database-view/DatabaseSOQLDetailPanel.ts
@@ -22,6 +22,7 @@ export class DatabaseSOQLDetailPanel extends LitElement {
         display: flex;
         overflow: hidden;
         width: 100%;
+        height: 100%;
       }
       call-stack {
         border-right: 2px solid var(--vscode-editorHoverWidget-border);
