+++ log-viewer/modules/components/database-view/DatabaseView.scss
@@ -1,6 +1,7 @@
 @import '../../datagrid/style/DataGrid.scss';
 
 .row__details-container {
+  height: 100%;
   padding: 4px 0px 4px 0px;
   background-color: var(--vscode-editorHoverWidget-background);
 }
